+++ workspace/packages/types/siyuan/window.d.ts
@@ -15,10 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { vitePreprocess } from '@sveltejs/vite-plugin-svelte'
-
-export default {
-	// Consult https://svelte.dev/docs#compile-time-svelte-preprocess
-	// for more information about preprocessors
-	preprocess: vitePreprocess(),
+export interface SiyuanWindow extends (Window & typeof globalThis) { 
+    siyuan: any;
 }
+