+++ workspace/plugins/jupyter-client/vite.config.ts
@@ -1,16 +1,16 @@
 /**
  * Copyright (C) 2023 Zuoqiu Yingyi
- *
+ * 
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
- *
+ * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Affero General Public License for more details.
- *
+ * 
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
@@ -32,7 +32,7 @@ export default defineConfig({
     ],
     build: {
         // REF https://cn.vitejs.dev/config/build-options.html#build-minify
-        minify: false,
+        minify: true,
         lib: {
             // Could also be a dictionary or array of multiple entry points
             entry: resolve(__dirname, "src/index.ts"),