+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - 添加附属仓库 `siyuan-plugin-jupyter-client` | Add sub-repository `siyuan-plugin-jupyter-client`.
 - 调整构建工具 | Adjust the build tools.
 - 初始化 `jupyter-client` 插件 | Initialize `jupyter-client` plugin.
+- 为 `jupyter-client` 插件提供 `svelte-i18n` 支持 | Provides `svelte-i18n` support for `jupyter-client` plugin.